# use Dask to parallelize document learning