--- a/packages/jupyter-ai/jupyter_ai/extension.py
+++ b/packages/jupyter-ai/jupyter_ai/extension.py
@@ -125,7 +125,6 @@ class AiExtension(ExtensionApp):
         # memory object used by the LM chain.
         self.settings["chat_history"] = []
 
-
         reply_queue = Queue()
         self.settings["reply_queue"] = reply_queue
 
@@ -154,7 +153,7 @@ class AiExtension(ExtensionApp):
         learn_actor = LearnActor.options(name=ACTOR_TYPE.LEARN.value).remote(
             reply_queue=reply_queue,
             log=self.log,
-            root_dir=self.serverapp.root_dir,
+            root_dir=self.serverapp.root_dir
         )
         ask_actor = AskActor.options(name=ACTOR_TYPE.ASK.value).remote(
             reply_queue=reply_queue,