--- a/packages/jupyter-ai/pyproject.toml
+++ b/packages/jupyter-ai/pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "tiktoken", # required for OpenAIEmbeddings
     "jupyter_ai_magics",
     "ray~=2.4.0", # Requires grpcio installation from conda
+    "dask[distributed]",
     "faiss-cpu", # Not distributed by official repo
     "wcmatch",
 ]