# Description

Uses Dask to parallelize document learning via `/learn`. The current approach sees roughly a 2-3x performance boost.

Before:
```
(LearnActor pid=13273) [/learn] Finished chunking documents. Time: 1308ms
(LearnActor pid=13273) [/learn] Completed. Time: 8117ms
```

After:
```
(LearnActor pid=33141) [/learn] Finished chunking documents. Time: 2685ms
(LearnActor pid=33141) [/learn] Finished computing embeddings. Time: 4571ms
(LearnActor pid=33141) [/learn] Complete. Time: 4573ms
```

# Notes

The majority of the time spent by Dask is actually from the overhead of spawning a process. The grey bars represent deserialization of pickled data in each worker process:

![Screen Shot 2023-05-22 at 12 07 49 PM](https://github.com/jupyterlab/jupyter-ai/assets/44106031/b86065a0-14d4-4c9a-9c72-b2f3f6a768ab)

There are some substantial improvements that can be made here.